app/requirements.txt
Flask==2.1.4
mysql-connector-python==8.0.23